Refactor: Style File 통합
- SignUp, Login, FirstLogin Style 파일에 겹치는 로직이 많아 통합
