+++ web/components/Modal/FirstLogin/FirstLoginPage/Page04.tsx
@@ -1,5 +1,5 @@
 import { MouseEventHandler } from 'react';
-import * as S from '../FirstLogin.style';
+import * as S from '../../Modal.style';
 import { Button, Input, Icon } from '../../../index';
 
 interface Props {
@@ -9,12 +9,11 @@ interface Props {
 
 const Page02 = ({ handleNextPage, handlePreviousPage }: Props) => {
   return (
-    <S.Container>
+    <>
       <S.PreviousButton onClick={handlePreviousPage}>
         <Icon name="arrowLeft" size={30} />
       </S.PreviousButton>
       <S.Title>
-        <br />
         Haeyum님만의
         <br />
         <S.MainText>프로필 사진</S.MainText>을 설정해 주세요!
@@ -28,7 +27,7 @@ const Page02 = ({ handleNextPage, handlePreviousPage }: Props) => {
           다음 &gt;
         </Button>
       </S.ButtonContainer>
-    </S.Container>
+    </>
   );
 };
